# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API ProviEmplea EVA3",
        "description": "API para gestión de candidatos con CV ciego, empresas, relación empresa-candidato y salud de la API.\n\n**Rate Limiting:** 60 requests/minuto por IP.\n**Caché:** Endpoints GET cacheados por 5 minutos.\n**Tiempo de respuesta objetivo:** < 200ms.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor local Docker"
        }
    ],
    "paths": {
...
        "/empresa-candidatos": {
            "get": {
                "tags": [
                    "Empresa-Candidatos"
                ],
                "summary": "Listar relaciones empresa-candidato",
                "description": "Obtiene todas las relaciones entre empresas y candidatos.",
                "operationId": "5041086b9f6cea43c5c3ecc7d4fcad80",
                "responses": {
                    "200": {
                        "description": "Lista de relaciones obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/EmpresaCandidato"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Empresa-Candidatos"
                ],
                "summary": "Crear relación empresa-candidato",
                "description": "Registra que una empresa está interesada en un candidato para un cargo específico.",
                "operationId": "548260aac9cf39272d4e433d93673018",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmpresaCandidato"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Relación creada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmpresaCandidato"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/empresa-candidatos/{id}": {
            "get": {
                "tags": [
                    "Empresa-Candidatos"
                ],
                "summary": "Obtener relación por ID",
                "description": "Obtiene una relación específica entre empresa y candidato.",
                "operationId": "caabd2ed047a7f2535456daf75a2f9ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la relación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Relación encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmpresaCandidato"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Relación no encontrada"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "put": {
                "tags": [
                    "Empresa-Candidatos"
                ],
                "summary": "Actualizar relación empresa-candidato",
                "description": "Actualiza el cargo buscado, estado del proceso u observación de la relación.",
                "operationId": "9b819881a8f32c4984c174e5af71f17c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la relación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmpresaCandidato"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Relación actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmpresaCandidato"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Relación no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Empresa-Candidatos"
                ],
                "summary": "Eliminar relación empresa-candidato",
                "description": "Elimina una relación entre empresa y candidato.",
                "operationId": "73d650e610a2275ff43f3f6222b8b85d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la relación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Relación eliminada correctamente"
                    },
                    "404": {
                        "description": "Relación no encontrada"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EmpresaCandidato": {
                "title": "Empresa Candidato",
                "required": [
                    "empresa_id",
                    "candidato_id",
                    "cargo_buscado"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "empresa_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "candidato_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "cargo_buscado": {
                        "type": "string",
                        "example": "Backend Developer"
                    },
                    "estado_proceso": {
                        "type": "string",
                        "example": "Pendiente"
                    },
                    "observacion": {
                        "type": "string",
                        "example": "Perfil interesante para entrevista"
                    }
                },
                "type": "object"
            }
        }
    },
...
}